backend/requirements.txt
Django>=4.1
djangorestframework
gunicorn
requests
python-dateutil
pytz
whitenoise
